==> 3_Proportion-Chart-together.R <==
#Gráfico de proporção - amostras juntas

new_dataset <- read.csv("new_output.csv", sep=";")
grafico <- ggplot(new_dataset) +
  aes(x = my_class, fill = variable, weight = frequencia) +
  geom_bar(position = "fill") +
  #geom_text(aes(label = paste0(round(prop * 100), "%"), y = prop), position = position_fill(vjust = 0.5)) +
  scale_x_discrete(labels = row.names(new_dataset)) +
  labs(x = "Amostras", y = "Frequência", fill = "Organismos mais relevantes", title = "Resultados RPIP Lote004 - Reads Totais") +
  theme_bw() +
  scale_fill_manual(values = c("Achromobacter denitrificans"= "#be29ec",
                               "Bartonella henselae"= "#be29ec",
                               "Cardiobacterium hominis"= "#be29ec",
                               "Elizabethkingia meningoseptica"= "#be29ec",
                               "Haemophilus influenzae"= "#be29ec",
                               "Mycobacterium gordonae"= "#be29ec",
                               "Achromobacter xylosoxidans"= "#be29ec",
                               "Bartonella quintana"= "#be29ec",
                               "Cardiobacterium valvarum"= "#be29ec",
                               "Enterobacter cloacae complex"= "#be29ec",
                               "Haemophilus parahaemolyticus"= "#be29ec",
                               "Mycobacterium kansasii"= "#be29ec",
                               "Acinetobacter baumannii"= "#be29ec",
                               "Bordetella bronchiseptica"= "#be29ec",
                               "Chlamydia pneumoniae"= "#be29ec",
                               "Enterococcus faecalis"= "#be29ec",
                               "Haemophilus parainfluenzae"= "#be29ec",
                               "Mycobacterium malmoense"= "#be29ec",
                               "Acinetobacter lwoffii"= "#be29ec",
                               "Bordetella hinzii"= "#be29ec",
                               "Chlamydia psittaci"= "#be29ec",
                               "Enterococcus faecium"= "#be29ec",
                               "Haemophilus pittmaniae"= "#be29ec",
                               "Mycobacterium parascrofulaceum"= "#be29ec",
                               "Acinetobacter nosocomialis"= "#be29ec",
                               "Bordetella holmesii"= "#be29ec",
                               "Chlamydia trachomatis"= "#be29ec",
                               "Escherichia coli"= "#be29ec",
                               "Hafnia alvei"= "#be29ec",
                               "Mycobacterium scrofulaceum"= "#be29ec",
                               "Acinetobacter pittii"= "#be29ec",
                               "Bordetella parapertussis"= "#be29ec",
                               "Chromobacterium violaceum"= "#be29ec",
                               "Eubacterium brachy"= "#be29ec",
                               "Klebsiella variicola"= "#be29ec",
                               "Mycobacterium szulgai"= "#be29ec",
                               "Actinomyces graevenitzii"= "#be29ec",
                               "Bordetella pertussis"= "#be29ec",
                               "Citrobacter freundii"= "#be29ec",
                               "Eubacterium limosum"= "#be29ec",
                               "Kytococcus sedentarius"= "#be29ec",
                               "Mycobacterium tuberculosis"= "#be29ec",
                               "Actinomyces israelii"= "#be29ec",
                               "Bordetella petrii"= "#be29ec",
                               "Citrobacter koseri"= "#be29ec",
                               "Eubacterium nodatum"= "#be29ec",
                               "Leclercia adecarboxylata"= "#be29ec",
                               "Mycobacterium xenopi"= "#be29ec",
                               "Actinomyces meyeri"= "#be29ec",
                               "Brucella abortus"= "#be29ec",
                               "Corynebacterium diphtheriae"= "#be29ec",
                               "Finegoldia magna"= "#be29ec",
                               "Legionella anisa"= "#be29ec",
                               "Mycobacteroides abscessus"= "#be29ec",
                               "Mycobacterium abscessus"= "#be29ec",
                               "Actinomyces naeslundii"= "#be29ec",
                               "Brucella canis"= "#be29ec",
                               "Corynebacterium jeikeium"= "#be29ec",
                               "Francisella tularensis"= "#be29ec",
                               "Legionella feeleii"= "#be29ec",
                               "Mycobacteroides chelonae"= "#be29ec",
                               "Mycobacterium chelonae"= "#be29ec",
                               "Actinomyces odontolyticus"= "#be29ec",
                               "Brucella melitensis"= "#be29ec",
                               "Corynebacterium propinquum"= "#be29ec",
                               "Fusobacterium necrophorum"= "#be29ec",
                               "Legionella longbeachae"= "#be29ec",
                               "Mycobacteroides immunogenum"= "#be29ec",
                               "Mycobacterium immunogenum"= "#be29ec",
                               "Aeromonas caviae"= "#be29ec",
                               "Brucella suis"= "#be29ec",
                               "Corynebacterium pseudodiphtheriticum"= "#be29ec",
                               "Fusobacterium nucleatum"= "#be29ec",
                               "Legionella maceachernii"= "#be29ec",
                               "Mycoplasma pneumoniae"= "#be29ec",
                               "Aeromonas hydrophila"= "#be29ec",
                               "Burkholderia cepacia complex"= "#be29ec",
                               "Corynebacterium pseudotuberculosis"= "#be29ec",
                               "Gemella haemolysans"= "#be29ec",
                               "Legionella pneumophila"= "#be29ec",
                               "Neisseria flavescens"= "#be29ec",
                               "Aeromonas sobria"= "#be29ec",
                               "Burkholderia gladioli"= "#be29ec",
                               "Corynebacterium striatum"= "#be29ec",
                               "Gemella morbillorum"= "#be29ec",
                               "Legionella wadsworthii"= "#be29ec",
                               "Neisseria lactamica"= "#be29ec",
                               "Aeromonas veronii"= "#be29ec",
                               "Burkholderia glumae"= "#be29ec",
                               "Corynebacterium ulcerans"= "#be29ec",
                               "Gordonia araii"= "#be29ec",
                               "Leptospira interrogans"= "#be29ec",
                               "Neisseria meningitidis"= "#be29ec",
                               "Aggregatibacter actinomycetemcomitans"= "#be29ec",
                               "Burkholderia mallei"= "#be29ec",
                               "Coxiella burnetii"= "#be29ec",
                               "Gordonia bronchialis"= "#be29ec",
                               "Leptotrichia buccalis"= "#be29ec",
                               "Neisseria mucosa"= "#be29ec",
                               "Aggregatibacter aphrophilus"= "#be29ec",
                               "Burkholderia pseudomallei"= "#be29ec",
                               "Cronobacter sakazakii"= "#be29ec",
                               "Haemophilus haemolyticus"= "#be29ec",
                               "Listeria monocytogenes"= "#be29ec",
                               "Nocardia abscessus"= "#be29ec",
                               "Arcanobacterium haemolyticum"= "#be29ec",
                               "Burkholderia thailandensis"= "#be29ec",
                               "Delftia acidovorans"= "#be29ec",
                               "Kingella kingae"= "#be29ec",
                               "Moraxella catarrhalis"= "#be29ec",
                               "Nocardia arthritidis"= "#be29ec",
                               "Bacillus anthracis"= "#be29ec",
                               "Campylobacter concisus"= "#be29ec",
                               "Dialister pneumosintes"= "#be29ec",
                               "Klebsiella aerogenes (Enterobacter aerogenes)"= "#be29ec",
                               "Moraxella osloensis"= "#be29ec",
                               "Nocardia beijingensis"= "#be29ec",
                               "Bacillus cereus"= "#be29ec",
                               "Capnocytophaga gingivalis"= "#be29ec",
                               "Dolosigranulum pigrum"= "#be29ec",
                               "Klebsiella oxytoca"= "#be29ec",
                               "Morganella morganii"= "#be29ec",
                               "Nocardia brasiliensis"= "#be29ec",
                               "Bacillus thuringiensis"= "#be29ec",
                               "Capnocytophaga leadbetteri"= "#be29ec",
                               "Eikenella corrodens"= "#be29ec",
                               "Klebsiella pneumoniae"= "#be29ec",
                               "Mycobacterium avium complex"= "#be29ec",
                               "Nocardia cyriacigeorgica"= "#be29ec",
                               "Bacteroides fragilis"= "#be29ec",
                               "Capnocytophaga sputigena"= "#be29ec",
                               "Elizabethkingia anophelis"= "#be29ec",
                               "Klebsiella quasipneumoniae"= "#be29ec",
                               "Mycobacterium fortuitum"= "#be29ec",
                               "Nocardia farcinica"= "#be29ec",
                               "Nocardia nova"= "#be29ec",
                               "Pediococcus acidilactici"= "#be29ec",
                               "Pseudomonas aeruginosa"= "#be29ec",
                               "Serratia marcescens"= "#be29ec",
                               "Streptococcus intermedius"= "#be29ec",
                               "Ureaplasma urealyticum"= "#be29ec",
                               "Nocardia otitidiscaviarum"= "#be29ec",
                               "Peptostreptococcus anaerobius"= "#be29ec",
                               "Pseudomonas fluorescens"= "#be29ec",
                               "Shewanella putrefaciens"= "#be29ec",
                               "Streptococcus mitis"= "#be29ec",
                               "Veilonella parvula"= "#be29ec",
                               "Nocardia transvalensis"= "#be29ec",
                               "Prevotella buccae"= "#be29ec",
                               "Pseudomonas stutzeri"= "#be29ec",
                               "Slackia exigua"= "#be29ec",
                               "Streptococcus pneumoniae"= "#be29ec",
                               "Williamsia muralis"= "#be29ec",
                               "Nocardia veterana"= "#be29ec",
                               "Prevotella intermedia"= "#be29ec",
                               "Ralstonia pickettii"= "#be29ec",
                               "Sphingomonas paucimobilis"= "#be29ec",
                               "Streptococcus pyogenes"= "#be29ec",
                               "Yersinia enterocolitica"= "#be29ec",
                               "Ochrobactrum anthropi"= "#be29ec",
                               "Prevotella melaninogenica"= "#be29ec",
                               "Raoultella ornithinolytica"= "#be29ec",
                               "Staphylococcus aureus"= "#be29ec",
                               "Tatlockia micdadei (Legionella micdadei)"= "#be29ec",
                               "Yersinia pestis"= "#be29ec",
                               "Orientia tsutsugamushi"= "#be29ec",
                               "Prevotella pleuritidis"= "#be29ec",
                               "Raoultella planticola"= "#be29ec",
                               "Stenotrophomonas maltophilia"= "#be29ec",
                               "Treponema denticola"= "#be29ec",
                               "Pandoraea pulmonicola"= "#be29ec",
                               "Proteus mirabilis"= "#be29ec",
                               "Rhodococcus hoagii"= "#be29ec",
                               "Streptococcus agalactiae"= "#be29ec",
                               "Tropheryma whipplei"= "#be29ec",
                               "Pantoea agglomerans"= "#be29ec",
                               "Proteus penneri"= "#be29ec",
                               "Rickettsia rickettsii"= "#be29ec",
                               "Streptococcus anginosus"= "#be29ec",
                               "Tsukamurella pulmonis"= "#be29ec",
                               "Parvimonas micra"= "#be29ec",
                               "Proteus vulgaris"= "#be29ec",
                               "Rothia mucilaginosa"= "#be29ec",
                               "Streptococcus constellatus"= "#be29ec",
                               "Tsukamurella tyrosinosolvens"= "#be29ec",
                               "Pasteurella multocida"= "#be29ec",
                               "Providencia stuartii"= "#be29ec",
                               "Salmonella enterica"= "#be29ec",
                               "Streptococcus dysgalactiae"= "#be29ec",
                               "Ureaplasma parvum"= "#be29ec",
                               "Coxsackievirus A"= "#ff4d00",
                               "HHV6"= "#ff4d00",
                               "Human Coronavirus NL63"= "#ff4d00",
                               "Human parechovirus"= "#ff4d00",
                               "Influenza C virus"= "#ff4d00",
                               "Rhinovirus A"= "#ff4d00",
                               "Coxsackievirus B"= "#ff4d00",
                               "Human adenovirus B"= "#ff4d00",
                               "Human Coronavirus OC43"= "#ff4d00",
                               "Influenza A virus"= "#ff4d00",
                               "Measles Virus"= "#ff4d00",
                               "Rhinovirus B"= "#ff4d00",
                               "Cytomegalovirus (CMV)"= "#ff4d00",
                               "Human adenovirus C"= "#ff4d00",
                               "Human metapneumovirus"= "#ff4d00",
                               "MERS coronavirus"= "#ff4d00",
                               "Rhinovirus C"= "#ff4d00",
                               "EBV"= "#ff4d00",
                               "Human adenovirus E"= "#ff4d00",
                               "Human parainfluenza virus 1"= "#ff4d00",
                               "Mumps virus"= "#ff4d00",
                               "Rubella virus"= "#ff4d00",
                               "Enterovirus A71"= "#ff4d00",
                               "Human bocavirus 1"= "#ff4d00",
                               "Human parainfluenza virus 2"= "#ff4d00",
                               "Influenza A virus (H7N9)"= "#ff4d00",
                               "Parvovirus B19"= "#ff4d00",
                               "SARS coronavirus"= "#ff4d00",
                               "Enterovirus D68"= "#ff4d00",
                               "Human Coronavirus 229E"= "#ff4d00",
                               "Human parainfluenza virus 3"= "#ff4d00",
                               "Respiratory Syncytial Virus A"= "#ff4d00",
                               "severe acute respiratory syndrome coronavirus 2"= "#ff4d00",
                               "Herpes simplex virus 1 (HSV-1)"= "#ff4d00",
                               "Human Coronavirus HKU1"= "#ff4d00",
                               "Human parainfluenza virus 4"= "#ff4d00",
                               "Influenza B virus"= "#ff4d00",
                               "Respiratory Syncytial Virus B"= "#ff4d00",
                               "Varicella-zoster virus (HHV-3)"= "#ff4d00",
                               "Alternaria alternata"= "#006666",
                               "Blastomyces dermatitidis"= "#006666",
                               "Curvularia lunata"= "#006666",
                               "Lomentospora prolificans (Scedosporium prolificans)"= "#006666",
                               "Pneumocystis jirovecii"= "#006666",
                               "Sarocladium kiliense (Acremonium kiliense)"= "#006666",
                               "Alternaria infectoria"= "#006666",
                               "Candida auris"= "#006666",
                               "Exophiala dermatitidis"= "#006666",
                               "Microascus cinereus"= "#006666",
                               "Scopulariopsis cinereus"= "#006666",
                               "Purpureocillium lilacinum"= "#006666",
                               "Scedosporium apiospermum"= "#006666",
                               "Apophysomyces elegans"= "#006666",
                               "Cladophialophora bantiana"= "#006666",
                               "Fusarium oxysporum"= "#006666",
                               "Microascus cirrosus (Scopulariopsis paisii)"= "#006666",
                               "Rasamsonia aegroticola"= "#006666",
                               "Schizophyllum commune"= "#006666",
                               "Aspergillus flavus"= "#006666",
                               "Coccidioides immitis"= "#006666",
                               "Fusarium proliferatum"= "#006666",
                               "Microasucus paisii"= "#006666",
                               "Scopulariopsis brumptii"= "#006666",
                               "Rasamsonia argillacea"= "#006666",
                               "Scopulariopsis brevicaulis"= "#006666",
                               "Aspergillus fumigatus"= "#006666",
                               "Coccidioides posadasii"= "#006666",
                               "Fusarium solani"= "#006666",
                               "Mucor circinelloides"= "#006666",
                               "Rhizomucor pusillus"= "#006666",
                               "Sporothrix schenckii"= "#006666",
                               "Aspergillus nidulans"= "#006666",
                               "Cryptococcus gattii"= "#006666",
                               "Fusarium verticillioides"= "#006666",
                               "Mucor indicus"= "#006666",
                               "Rhizopus azygosporus"= "#006666",
                               "Syncephalastrum racemosum"= "#006666",
                               "Aspergillus niger"= "#006666",
                               "Cryptococcus neoformans"= "#006666",
                               "Histoplasma capsulatum"= "#006666",
                               "Mucor racemosus"= "#006666",
                               "Rhizopus microsporus"= "#006666",
                               "Talaromyces marneffei"= "#006666",
                               "Aspergillus terreus"= "#006666",
                               "Cunninghamella bertholletiae"= "#006666",
                               "Lichtheimia corymbifera"= "#006666",
                               "Paecilomyces variotii"= "#006666",
                               "Rhizopus oryzae"= "#006666",
                               "Trichosporon asahii"= "#006666",
                               "Aspergillus versicolor"= "#006666",
                               "Curvularia geniculata"= "#006666",
                               "Lichtheimia ramosa"= "#006666",
                               "Paracoccidioides brasiliensis"= "#006666",
                               "Saksenaea vasiformis"= "#006666")) +
                               
  theme(
    panel.border = element_rect(color = "black", fill = NA, size = 1),
    axis.text = element_text(face = "bold", size = 12),
    axis.title = element_text(face = "bold", size = 16),
    legend.title = element_text(size = 10, face = "bold"),
    legend.text = element_text(size = 8, face = "bold"),
    axis.text.x = element_text(size = 8),
    axis.text.y = element_text(size = 8),
    axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0)))


ggsave("grafico.jpg", plot = grafico, width = 16, height = 8, dpi = 600)
